Detect ticker to classification module
